+++ src/lib/constants/index.ts
@@ -1,5 +1,4 @@
 // Constant exports
 // Placeholder files will be implemented in future tasks
-// export * from './routes'
 // export * from './animations'
 // export * from './seo'